Allow setting minPlayers & close websocket on game start + other fixes

+++ Extensions/Multiplayer/multiplayerobjectruntimebehavior.ts
@@ -67,9 +67,6 @@ namespace gdjs {
     _destroyInstanceTimeoutId: NodeJS.Timeout | null = null;
     _timeBeforeDestroyingObjectWithoutNetworkIdInMs = 500;
 
-    // Boolean to help with the destruction of the object.
-    _shouldDeleteObject = false;
-
     constructor(
       instanceContainer: gdjs.RuntimeInstanceContainer,
       behaviorData,
@@ -86,19 +83,6 @@ namespace gdjs {
           ? window.performance.now.bind(window.performance)
           : Date.now;
 
-      // If the object is assigned to a player that does not exist, mark it for destruction before the next tick.
-      // We cannot destroy it immediately as the object is being created.
-      if (
-        this.playerNumber !== 0 && // Host is always connected.
-        !gdjs.multiplayerMessageManager.isPlayerConnected(this.playerNumber)
-      ) {
-        logger.info(
-          `Player number ${this.playerNumber} does not exist in the lobby at the moment. Marking the object for destruction.`
-        );
-        this._shouldDeleteObject = true;
-        return;
-      }
-
       // When a synchronized object is created, we assume it will be assigned a networkId quickly if:
       // - It is a new object created by the current player. -> will be assigned a networkId when sending the update message.
       // - It is an object created by another player. -> will be assigned a networkId when receiving the update message.
@@ -107,18 +91,13 @@ namespace gdjs {
       // To handle this case and avoid having an object not synchronized, we set a timeout to destroy the object
       // if it has not been assigned a networkId after a short delay.
       this._destroyInstanceTimeoutId = setTimeout(() => {
-        if (!owner.networkId) {
+        if (!owner.networkId && gdjs.multiplayer.isLobbyGameRunning()) {
           logger.info(
-            `Object ${owner.getName()} has not been assigned a networkId after a short delay, destroying it.`
+            `Lobby game is running and object ${owner.getName()} has not been assigned a networkId after a short delay, destroying it.`
           );
           owner.deleteFromScene(instanceContainer);
         }
       }, this._timeBeforeDestroyingObjectWithoutNetworkIdInMs);
-
-      logger.info(
-        'Created MultiplayerObjectRuntimeBehavior object with name ' +
-          owner.getName()
-      );
     }
 
     sendDataToPeersWithIncreasedClock(messageName: string, data: Object) {
@@ -263,8 +242,19 @@ namespace gdjs {
     }
 
     doStepPostEvents() {
-      // Before doing anything, check if the object is marked for destruction.
-      if (this._shouldDeleteObject) {
+      // Before doing anything, check if the game is running, if not, return.
+      if (!gdjs.multiplayer.isLobbyGameRunning()) {
+        return;
+      }
+
+      // If game is running and object belong to a player that is not connected, destroy the object.
+      if (
+        this.playerNumber !== 0 && // Host is always connected.
+        !gdjs.multiplayerMessageManager.isPlayerConnected(this.playerNumber)
+      ) {
+        logger.info(
+          `Player number ${this.playerNumber} does not exist in the lobby at the moment. Destroying the object.`
+        );
         this.owner.deleteFromScene(this.owner.getInstanceContainer());
         return;
       }
@@ -407,6 +397,11 @@ namespace gdjs {
         this._destroyInstanceTimeoutId = null;
       }
 
+      // If the lobby game is not running, just return here.
+      if (!gdjs.multiplayer.isLobbyGameRunning()) {
+        return;
+      }
+
       // For desruction of objects, we allow the host to destroy the object even if it is not the owner.
       // This is particularly helpful when a player disconnects, so the host can destroy the object they were owning.
       if (!this.isOwnerAsPlayerOrHost() && !gdjs.multiplayer.isPlayerHost()) {
@@ -496,28 +491,24 @@ namespace gdjs {
         );
         return;
       }
-      if (
-        newPlayerNumber !== 0 && // Host is always connected.
-        !gdjs.multiplayerMessageManager.isPlayerConnected(newPlayerNumber)
-      ) {
-        logger.info(
-          `Player number ${newPlayerNumber} does not exist in the lobby at the moment. Destroying the object as it will not be synchronized.`
-        );
-        if (this._destroyInstanceTimeoutId) {
-          clearTimeout(this._destroyInstanceTimeoutId);
-          this._destroyInstanceTimeoutId = null;
-        }
-        this.owner.deleteFromScene(this.owner.getInstanceContainer());
+
+      // Update the ownership locally, so the object can be used immediately.
+      // This is a prediction to allow snappy interactions.
+      // If we are player 1 or host, we will have the ownership immediately anyway.
+      // If we are another player, we will have the ownership as soon as the host acknowledges the change.
+      // If the host does not send an acknowledgment, we will revert the ownership.
+      this.playerNumber = newPlayerNumber;
+
+      // If the lobby game is not running, just return here.
+      if (!gdjs.multiplayer.isLobbyGameRunning()) {
         return;
       }
 
       let instanceNetworkId = this.owner.networkId;
-
       if (!instanceNetworkId) {
         logger.info(
           'Object has no networkId, we change the ownership locally, but it will not be synchronized yet if we are not the owner.'
         );
-        this.playerNumber = newPlayerNumber;
         if (newPlayerNumber !== gdjs.multiplayer.getPlayerNumber()) {
           // If we are not the new owner, we should not send a message to the host to change the ownership.
           // Just return and wait to receive an update message to reconcile this object.
@@ -576,13 +567,6 @@ namespace gdjs {
         logger.info('Sending change owner message', messageName);
         this.sendDataToPeersWithIncreasedClock(messageName, messageData);
       }
-
-      // We also update the ownership locally, so the object can be used immediately.
-      // This is a prediction to allow snappy interactions.
-      // If we are player 1 or host, we will have the ownership immediately anyway.
-      // If we are another player, we will have the ownership as soon as the host acknowledges the change.
-      // If the host does not send an acknowledgment, we will revert the ownership.
-      this.playerNumber = newPlayerNumber;
 
       // If we are the new owner, also send directly an update of the position,
       // so that the object is immediately moved on the screen and we don't wait for the next tick.
